Refactor close map popup after setFilters
